pyproject.toml
[project]
name = "ai-qa-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.56.0",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "jsonschema>=4.25.1",
    "langchain>=1.2.0",
    "langchain-core>=1.2.5",
    "langchain-google-genai>=4.1.2",
    "langchain-mcp-adapters>=0.2.1",
    "langdetect>=1.0.9",
    "mcp>=1.25.0",
    "pdfplumber>=0.11.8",
    "playwright>=1.57.0",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "sse-starlette>=2.1.0",
    "starlette>=0.41.0",
    "fastmcp>=2.14.2",
    "pytest-asyncio>=1.3.0",
]


[tool.setuptools]
packages = [
    "mcp_client",
]